@@ -16,8 +16,8 @@
     "@emotion/styled": "^11.13.0",
     "framer-motion": "^11.3.8",
     "next": "14.2.5",
-    "react": "^0.0.0-experimental-14a4699f-20240725",
-    "react-dom": "^0.0.0-experimental-14a4699f-20240725",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "react-icons": "^5.2.1"
   },
   "devDependencies": {
