@@ -1,9 +1,3 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-    experimental: {
-      appDir: true,
-    },
-  };
-  
-  module.exports = nextConfig;
-  
+const nextConfig = {};
+
+export default nextConfig;
